fixed mercurial PreReceiveRepositoryHooks
The problem seems to be that guice had multiple options for injecting
HgContext. HgContextProvider bound via Module and HgContext bound void
RequestScoped annotation. It looks like that Guice 4 injects randomly
the one or the other, in SCMv1 (Guice 3) everything works as expected.

To fix the problem we have created a new class annotated with
RequestScoped, which holds an instance of HgContext. This way only the
HgContextProvider is used for injection.