Add POC for diff against temporary merge result
This fails for files with merge conflicts, because results for these
files are not added to git and therefore there is no blob the diff
formatter can use to render. This has somehow to be fixed by making
jgit use the version from the file system. Taking a look at the
implementation of Git.clone() might help here.