Only use LF for line breaks in diffs
Git uses LF for line breaks in diffs, not CR or other delimiters. When
we are using other delimiters for diving diff output into lines, too,
we can get errors because diff lines can contain CRs. When we try to
split such lines, we get exceptions because these lines cannot be parsed