Fix migration of non-bare git repositories
During the migration of git repositories from v1 to v2, we have to
create an "scmm" config section with the repository id of the current
repository. If this does not happen, further write requests to this
repository will fail, because the hooks cannot determine the id.

However, the migration failed to write this configuration for non-bare
repositories. Therefore this fix checks beforehand, whether a '.git'
folder exists in the date directory. If this is the case, we assume that
this is a non-bare repository and write the config file inside this
folder.