Fix assumption
Sonar says:
AssertJ assertions allMatch and doesNotContains on an empty list always
returns true whatever the content of the predicate. Despite being
correct, you should make explicit if you expect an empty list or not, by
adding isEmpty()/isNotEmpty() before calling the assertion. It will
justify the useless predicate to improve clarity or increase the
reliability of the test.