Fixes NullPointerException
There may be v1 properties with key/value pairs, where the value is not
set. Such "pairs" led to a null pointer exception in the `map` function,
because no Optional could be created for the null value.
To solve this, we filter such properties without value in the beginning,
because these properties would create a null result both ways.