Fix scm home folder for integration tests
This variable should be absolute, because otherwise this may resolve
to different directories. For instance the web server resolves the
relative path to the root of the parent pom, while the dependency
plugin resolves it to the path of the scm-it pom. This led to the
error, that the scm-integration-test-plugin was not copied to the
correct plugin directory and therefore was not installed by the test
scm server.