Make request for feed more resilient
This commit tackles the following issue:
When the release feed takes a long time to read (however), the former
solution would propagete requests to the SCM-Manager to multiple
requests of the specified feed url. This can, in the worst case, cause
a drain of resources like request threads.

We fix this with two actions:
1. We wrap the request itself in an executor with a timeout,
2. We only trigger one request at a time, so that we will not flood
   the feed server with requests.