Suppress SonarQube warning after fix
The issue javasecurity:S5144 which was claimed by SonarQube is now fixed.
But SonarQube will find still find the issue, because we will send the request with user controller data, but we have now a secure check if the host is SCM-Manager.
So the issue javasecurity:S5144 is now a false-positive and will be suppressed.