Fix stacktrace in response for invalid url
This fixes responses with complete stack traces for
requests with invalid urls, for example such containing
backslash ('\') in the query parameter part (eg. q\=search).
In this case the response contains an error object due to
this error, and requesting the uri info would trigger the
same error a second time, only that now the exception mapper
would not catch the error again. So we check whether we have
an error object before trying to create an enricher context.