Fix hg fileview resolves for same path on file and directory (#1746)
If a file and a directory with the same name existed somewhere at the same level in a Mercurial repository, our logic in the fileview command failed to collect them. The parent of the file was mistaken for the entire file path, resulting in confusing errors that the file could not be found in the manifest. With this fix, file detection should now be more secure than before.

Co-authored-by: Sebastian Sdorra <sebastian.sdorra@cloudogu.com>