Do not redirect after data update (#1965)
Do not redirect to other page after data was simply updated. After removing this redirect we stumbled upon some bigger issue with stale data inside our forms. This led to conflicts (Concurrent Modification Exceptions) if data was updated multiple times without reloading the page because of the last modified date of the single entities. This issue was hidden through the redirect and we fixed this by refactoring the form to functional components and simplify the state handling.