remove module from queue before it is re-defined
If a queued module is only removed from the queue upon successfully resolving all its dependencies,
it could fail multiple times, causing multiple entries to appear in the queue.
This in turn causes the module to get executed multiple times.
The solution is to immediately remove the module when it is de-queued.
If it fails again, it is then re-added to the queue.