Use Java 11 as compile version (#2033)
This sets Java 11 for compilation. To do so, this has to use the gradle smp plugin with the pull request scm-manager/gradle-smp-plugin#16. If the smp plugin is not released as version 0.10.4, the version has to be corrected for org.scm-manager.smpin the various build.gradle files.

It might be necessary to delete all build folders when switching between Java 8 and Java 11:

rm -rf build */build scm-plugins/*/build

Co-authored-by: Eduard Heimbuch <eduard.heimbuch@cloudogu.com>